corrected migration for ai tutor interaction feedback

+++ dashboard/db/migrate/20240403074309_create_ai_tutor_interaction_feedbacks.rb
@@ -2,7 +2,7 @@ class CreateAiTutorInteractionFeedbacks < ActiveRecord::Migration[6.1]
   def change
     create_table :ai_tutor_interaction_feedbacks do |t|
       t.bigint :ai_tutor_interaction_id, null: false
-      t.bigint :user_id, null: false
+      t.integer :user_id, null: false
       t.boolean :thumbs_up
       t.boolean :thumbs_down
